requirements.txt
chromadb==0.4.17
anthropic==0.49.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
requests==2.31.0
tqdm==4.66.1
huggingface-hub==0.16.4
torch>=1.6.0 